Move some app content to the root of the store
