+++ web/app/containers/product-list/selectors.js
@@ -7,7 +7,7 @@ import {createSelector} from 'reselect'
 import {createGetSelector, createHasSelector} from 'reselect-immutable-helpers'
 import {getUi, getCategories} from '../../store/selectors'
 import {getSelectedCategory, getCategoryProducts} from '../../store/categories/selectors'
-import {getCurrentPathKey} from '../app/selectors'
+import {getCurrentPathKey} from '../../store/app/selectors'
 import {byFilters} from '../../utils/filter-utils'
 import {sortLib} from '../../utils/sort-utils'
