+++ web/app/store/app/constants.js
@@ -5,10 +5,4 @@
 /**
  * This file contains constants for use in the application.
  */
-export const PLACEHOLDER = 'placeholder'
-
-// This is the asset url we use to check if we're offline when making requests
-export const OFFLINE_ASSET_URL = '//online.mobify.net/online.json'
-
-// This is the Location page's url
-export const LOCATION_URL = '//locations.merlinspotions.com'
+export const CURRENT_URL = 'currentURL'
