+++ web/app/store/categories/actions.js
@@ -2,7 +2,7 @@
 /* Copyright (c) 2017 Mobify Research & Development Inc. All rights reserved. */
 /* * *  *  * *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  *  * */
 import {createAction} from 'progressive-web-sdk/dist/utils/action-creation'
-import {getCurrentPathKey} from '../../containers/app/selectors'
+import {getCurrentPathKey} from '../app/selectors'
 import {getSelectedCategory} from './selectors'
 
 export const changeFilter = createAction('Change Filter')
