+++ web/app/store/categories/selectors.js
@@ -6,7 +6,7 @@ import Immutable from 'immutable'
 import {createSelector} from 'reselect'
 import {createGetSelector} from 'reselect-immutable-helpers'
 import {getCategories, getProducts} from '../selectors'
-import {getCurrentPathKey} from '../../containers/app/selectors'
+import {getCurrentPathKey} from '../app/selectors'
 import {PLACEHOLDER} from '../../containers/app/constants'
 
 export const getSelectedCategory = createGetSelector(
