+++ web/app/store/selectors.js
@@ -10,3 +10,4 @@ export const getCart = ({cart}) => cart
 export const getNotifications = ({notifications}) => notifications
 export const getCheckout = ({checkout}) => checkout
 export const getForm = ({form}) => form
+export const getApp = ({app}) => app
