+++ web/app/template.jsx
@@ -1,6 +1,6 @@
 import React, {PropTypes} from 'react'
 import {connect} from 'react-redux'
-import {onRouteChanged} from './containers/app/actions'
+import {onRouteChanged} from './store/app/actions'
 import {removeAllNotifications} from 'progressive-web-sdk/dist/store/notifications/actions'
 import {incrementPageCount} from 'progressive-web-sdk/dist/store/push-messaging/actions'
